Define the module to be loaded in __init__.py
Since some modules may depend on each other and bsmedit may not be able to load
them in proper order. It tells the bsmedit to only load the specific modules
and ignore all the others. Otherwise, bsmedit may try to load all the modules
in that folder in random order. It also makes the loading slightly faster.