Fix a bug in closing (not destroying) the panel.
A flag was set on the AuiPanelInfo of each panel to indicate that the panel
should not be destroyed when clicking the close button. However, AUI library
may re-generate the AuiPanelInfo for the panel when the panel is added to a
notebook, and our flag may be lost during that process. Thus, add the flag
directly to the panel object (window).

Show icon on the notebook page.