Remove the top level simulation functions.
The top level function may cause conflict with the function in other packages
(e.g., step). One solution is to use "import pysim", instead of "from pysim
import *". However, in this case, the function call may look like "pysim.read",
which is not significantly simpler than "s.read". To minimize the confusion, we
delete most top level functions. So, for example, instead of "read()", now it
only supports "s.read()", where s is the handle to a simulation.

Add shortcut to have access to signals. For example,
a = s['signal'] <--> a = s.read('signal')
s['signal'] = a <--> s.write({'signal':a})

Update the name of some interface functions.