Update the SystemC interface to read the float/int/string values separately.
The value returned from the simulation was always a string for simplicity. The
drawback is that the float number is not accurate. So now the float/int/string
is returned according to the object type.

Add a 'get_status' command to retrieve the simulation status.

Add a 'wait_until_simulation_paused' command to wait for the command to pause
without blocking the GUI.